poker-backend/pyproject.toml
[tool.poetry]
name = "poker-backend"
version = "0.1.0"
description = "Backend for simplified poker exercise (FastAPI + asyncpg + pokerkit)"
authors = ["Eyerusalem kidanejerry523@gmail.com"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95"
uvicorn = {extras = ["standard"], version = "^0.21"}
asyncpg = "^0.30.0"
pydantic = "^1.10"
pokerkit = "^0.6"      # install the pokerkit library
python-dotenv = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
httpx = "^0.24"
pytest-cov = "^4.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
